+++ src/indexes/index-stream.ts
@@ -1,10 +1,10 @@
-import { Transform } from 'stream';
+import { Transform, TransformCallback } from 'stream';
 
-import { promiseOrCallback } from '../promise-or-callback';
+import { IndexingResult } from 'azure-search-types';
 import { SearchRequester } from '../search-requester';
 import { SearchOptions, SearchRequest } from '../types';
 import { IndexBuffer } from './index-buffer';
-import { IndexingResults } from './search-index';
+import { IndexDocument, IndexingResults } from './search-index';
 
 export class IndexStream extends Transform {
   private buffer = new IndexBuffer(async (data) => {
@@ -14,17 +14,19 @@ export class IndexStream extends Transform {
       headers: { 'content-type': 'application/json' },
       body: data,
     });
-    // TODO: pass IndexingResults as the stream transform
+    return resp.result.value;
   });
 
   constructor(private requester: SearchRequester, private index: string, private options?: SearchOptions) {
     super({
       objectMode: true,
       async transform(chunk, enc, cb) {
-        await promiseOrCallback(() => self.buffer.add(chunk), cb);
+        // await promiseOrCallback(() => self.buffer.add(chunk), cb);
+        await self.process(cb, chunk);
       },
       async flush(cb) {
-        await promiseOrCallback(() => self.buffer.flush(), cb);
+        // await promiseOrCallback(() => self.buffer.flush(), cb);
+        await self.process(cb);
       },
     });
 
@@ -34,5 +36,19 @@ export class IndexStream extends Transform {
   private request<T>(req: SearchRequest<T>) {
     req.path = `/indexes/${this.index}${req.path}`;
     return this.requester.request<T>(req, this.options);
+  }
+
+  private async process(cb: TransformCallback, document?: IndexDocument & Document) {
+    let maybeResults: IndexingResult[];
+    let error: any = null;
+    try {
+      maybeResults = document
+        ? await this.buffer.add(document)
+        : await this.buffer.flush();
+    } catch (err) {
+      error = err;
+    }
+
+    cb(error, maybeResults);
   }
 }